requirements.txt
streamlit==1.28.0
supabase==1.0.3
pandas==2.1.1
plotly==5.17.0
python-dotenv==1.0.0
streamlit-calendar==0.4.0